Moved integration tests into separate unit tests project
