+++ Hrimsoft.SqlBulk.PostgreSql.IntegrationTests/TestModels/SimpleEntityProfile.cs
@@ -1,6 +1,6 @@
 using System.Data;
 
-namespace Hrimsoft.SqlBulk.PostgreSql.Tests.TestModels
+namespace Hrimsoft.SqlBulk.PostgreSql.IntegrationTests.TestModels
 {
     public class SimpleEntityProfile: EntityProfile
     {
